% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_Simulation_Output.R
\name{NFL_Playoff_Probabilities}
\alias{NFL_Playoff_Probabilities}
\title{Obtain simulated NFL playoff probabilities}
\usage{
NFL_Playoff_Probabilities(initialization_start, season_to_sim, week_from,
  cval = 40.29, nsim = 10000, updates = TRUE, plot = TRUE)
}
\arguments{
\item{initialization_start}{Numeric, the system start year for calculating NFL teams' Glicko ratings. Must be greater than or equal to 1922, due to data availability on Pro-Football-Reference. See notes for selection guidelines}

\item{season_to_sim}{Numeric, the year of the current NFL season}

\item{week_from}{Numeric, the current NFL_week}

\item{cval}{Numeric, c-value to be used for Glicko rating algorithm. Default is \code{40.29}}

\item{nsim}{Numeric, the number of simulated seasons to produce. Default is \code{10000}}

\item{updates}{Logical, should progress updates be printed as the process runs? Default is \code{TRUE}}

\item{plot}{Logical, should a plot of simulated playoff probabilities be produced as part of the process? Default is \code{TRUE}}
}
\value{
A data frame containing each teams' simulated overall playoff, division win, and wild card probabilties.
  If \code{plot=TRUE}, a \code{ggplot2} style plot of this data will be generated in the plot window.
}
\description{
Applies the Glicko rating system to simulate multiple possibilities for the 
remainder of the current NFL season, in an effort to estimate the likelihood
that each team reaches the playoffs at a given point in the season.
}
\details{
The \code{initialization_start} parameter controls the start date for Glicko ratings, and
is used as a starting time for ratings calculations; this means that it is considered
by the algorithm as a time at which all teams "enter the system" (and, specifically, are equal
in rating upon entry). Though this parameter can accept any value greater than or equal to 1922 and
less than the current year, it should be logically constrained to years that are defendable as 
both origin points and times at which all teams may/should be considered equal. Some reasonable 
selections may include (but are not limited to):

1. 1922 -- the earliest available year for data (the NFL was founded in 1920)

2. 1966 -- the first season for which a Super Bowl was played

3. 1970 -- the year of the AFL-NFL merger (beginning of the "modern era" in the NFL)


If an NFL week is only partially complete at the time of simulation, the \code{week_from}
parameter should still be the number of that partially completed week. For example, if all
but one game has been completed in week 11, set \code{week_from=11}. If no games have 
yet been played in week 12 but week 11 has been completed entirely, set \code{week_from=12}.
If even just one game has been completed in week 12, set \code{week_from=12}. Simulation will
not overwrite results of partially completed weeks, but instead simulate the games from
that week that have not been completed at the time of simulation.
}
\examples{
#playoff probability for the New York Jets after week 11 of the 2018 season
playoff_probs = NFL_Playoff_Probabilities(1970, 2018, 11)
playoff_probs[playoff_probs$team == "New York Jets",]

}
\seealso{
Dr. Mark Glickman's description of the Glicko Rating system: \code{http://www.glicko.net/glicko/glicko.pdf}
}
